Fix failing doctests

+++ skimage/segmentation/random_walker_segmentation.py
@@ -332,7 +332,7 @@ def random_walker(data, labels, beta=130, mode='bf', tol=1.e-3, copy=True,
            [1, 1, 1, 1, 1, 2, 2, 2, 1, 1],
            [1, 1, 1, 1, 1, 2, 2, 2, 1, 1],
            [1, 1, 1, 1, 1, 1, 1, 1, 1, 1],
-           [1, 1, 1, 1, 1, 1, 1, 1, 1, 1]])
+           [1, 1, 1, 1, 1, 1, 1, 1, 1, 1]], dtype=int32)
 
     """
     # Parse input data
